+++ pocx_plotter/src/plotter.rs
@@ -60,7 +60,7 @@ pub struct PlotterTask {
     pub seed: Option<[u8; 32]>,
     pub warps: Vec<u64>,
     pub number_of_plots: Vec<u64>,
-    pub compress: u32,
+    pub compress: u8,
     pub output_paths: Vec<String>,
     pub mem: String,
     pub cpu_threads: u8,
@@ -237,7 +237,7 @@ impl Plotter {
                 )
             })?;
 
-        let compression_multiplier = u64::pow(2, task.compress);
+        let compression_multiplier = u64::pow(2, task.compress as u32);
         let mem_plot = compression_multiplier.checked_mul(mem_write)
             .ok_or_else(|| PoCXPlotterError::Config(format!(
                 "Memory plot calculation overflow: compression value {} results in 2^{} = {} which causes overflow when multiplied by memory requirements",
@@ -286,7 +286,7 @@ impl Plotter {
                 "Insufficient host memory for plotting!\n\nRAM: Total={:.2} GiB, Available={:.2} GiB\nPlotter requirement: {:.2} GiB x{} (escalation)\nWriter requirement: {:.2} GiB x{} (escalation)\nGPU requirement: {:.2} GiB\nTOTAL requirement: {:.2} GiB",
                 sys.total_memory() as f64 / 1024.0 / 1024.0 / 1024.0,
                 sys.available_memory() as f64 / 1024.0 / 1024.0 / 1024.0,
-                (u64::pow(2, task.compress) * DIM * NONCE_SIZE) as f64 / 1024.0 / 1024.0 / 1024.0,
+                (u64::pow(2, task.compress as u32) * DIM * NONCE_SIZE) as f64 / 1024.0 / 1024.0 / 1024.0,
                 task.escalate,
                 WARP_SIZE as f64 / 1024.0 / 1024.0 / 1024.0,
                 task.escalate,
@@ -346,7 +346,7 @@ impl Plotter {
 
             println!(
                 "     Cache(Plotter)={:.2} GiB x{} (escalation), Cache(HDD)={:.2} GiB x{} (escalation) x{} (disks), Cache(GPU)={:.2} GiB,\n",
-                (u64::pow(2, task.compress) * DIM * NONCE_SIZE) as f64 / 1024.0 / 1024.0 / 1024.0,
+                (u64::pow(2, task.compress as u32) * DIM * NONCE_SIZE) as f64 / 1024.0 / 1024.0 / 1024.0,
                 task.escalate,
                 (DIM * NONCE_SIZE) as f64 / 1024.0 / 1024.0 / 1024.0,
                 task.escalate,
@@ -372,7 +372,7 @@ impl Plotter {
             );
             println!(
                 "Compression    : {:?}(X{:?})",
-                u64::pow(2, task.compress),
+                u64::pow(2, task.compress as u32),
                 task.compress
             );
             println!("Output path(s) : {:?}", task.output_paths);
